refactor(tests): split service_tests.rs under 2000 lines for linecheck-2000
Extracts trigger/snooze tests into service_trigger_tests.rs and logs
tests into service_logs_tests.rs so all .rs files stay under 2000 lines.
